% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{practice.function}
\alias{practice.function}
\title{Squares a number or a vector of numbers}
\usage{
practice.function(x)
}
\arguments{
\item{x}{a number or a numeric vector to be squared}
}
\value{
the square of x
}
\description{
Squares a number or a vector of numbers
}
\examples{
practice.function(10) #100
practice.function(c(2,4)) # c(4,16)
}
